+++ kuznyechik/src/fused_tables/consts.rs
@@ -1,6 +1,12 @@
 // TODO: use u128 tables after MSRV is bumped to 1.77 or higher
 #[repr(align(16))]
-pub struct Align16<T>(pub T);
+pub(crate) struct Align16<T>(pub T);
+
+impl<T> Align16<T> {
+    pub fn as_ptr(&self) -> *const u128 {
+        self as *const Self as *const u128
+    }
+}
 
 pub type Table = Align16<[u8; 16 * 4096]>;
